# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/tunnel/CreateTunnel": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tunnel"
                ],
                "summary": "用户申请创建通道",
                "parameters": [
                    {
                        "description": "用户ID，隧道参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateTunnel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "隧道创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/user/changePassword": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tunnel"
                ],
                "summary": "用户查询其拥有的所有通道",
                "parameters": [
                    {
                        "description": "用户ID",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.FindUserAllTunnels"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "隧道查询成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.CreateTunnel": {
            "type": "object",
            "properties": {
                "localip": {
                    "type": "string",
                    "example": "127.0.0.1"
                },
                "localport": {
                    "type": "integer",
                    "example": 22
                },
                "remoteip": {
                    "type": "string",
                    "example": "139.254.14.13"
                },
                "remoteport": {
                    "type": "integer",
                    "example": 23
                },
                "token": {
                    "type": "string",
                    "example": "123456"
                },
                "tunnelname": {
                    "type": "string",
                    "example": "ssh"
                },
                "tunnelpoint": {
                    "type": "string",
                    "example": "shanghai"
                },
                "tunneltype": {
                    "type": "string",
                    "example": "tcp"
                }
            }
        },
...
        "request.FindUserAllTunnels": {
            "type": "object",
            "properties": {
                "userId": {
                    "type": "integer"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}